refactor: helper function and id constants modules
