+++ constants/ids.py
@@ -8,7 +8,8 @@ US_CPI_DURABLES = "CUUR0000SAD"
 US_INF_EXPECTATIONS = "MICH"
 US_NONDURABLES_CONSUMPTION = "PCEND"
 US_DURABLES_CONSUMPTION = "PCEDG"
-US_SAVINGS = "PSAVERT"
+US_SAVINGS = "PMSAVE"
+US_SAVINGS_RATE = "PSAVERT"
 
 # * French data (INSEE API)
 # TODO
