+++ src/helpers.py
@@ -1,15 +1,10 @@
 """Cross-project helper functions"""
 
-import logging
 from typing import Dict, Generator, List
 
+import pandas as pd
 
-def define_other_module_log_level(level: str) -> None:
-    """Disable logger ouputs for other modules up to defined `level`"""
-    for log_name in logging.Logger.manager.loggerDict:
-        if log_name != "__name__":
-            log_level = getattr(logging, level.upper())
-            logging.getLogger(log_name).setLevel(log_level)
+from src import retrieve_data
 
 
 def nested_dict_values(nested_dict: Dict) -> Generator[any, any, any]:
@@ -28,3 +23,14 @@ def nested_list_values(nested_list: List[List[str]]) -> Generator[any, any, any]
             yield from nested_list_values(v)
         else:
             yield v
+
+
+def add_real_value_columns(
+    data: pd.DataFrame, nominal_columns: List[str], **kwargs
+) -> pd.DataFrame:
+    """Convert nominal to real values for each column in list"""
+    for col in nominal_columns:
+        data[f"real_{col}"] = retrieve_data.convert_column_to_real_value(
+            data=data, column=col, **kwargs
+        )
+    return data
