+++ src/regression.py
@@ -16,7 +16,7 @@ import statsmodels.api as sm
 import statsmodels.iolib.summary2
 
 from src import dwt
-from src.helpers import define_other_module_log_level
+from src.logging_helpers import define_other_module_log_level
 from src import retrieve_data
 
 # * Logging settings
