+++ src/retrieve_data.py
@@ -14,8 +14,8 @@ import seaborn as sns
 import requests
 import xmltodict
 
-from src.helpers import define_other_module_log_level
-from src import ids
+from src.logging_helpers import define_other_module_log_level
+from constants import ids
 
 # * Logging settings
 logger = logging.getLogger(__name__)
