+++ tests/test_descriptive_stats.py
@@ -13,8 +13,8 @@ import pandas as pd
 from scipy import stats
 
 from src import descriptive_stats
-from src.helpers import define_other_module_log_level
-from src import ids
+from src.logging_helpers import define_other_module_log_level
+from constants import ids
 from src import retrieve_data
 
 # * Logging settings
