+++ tests/test_regression.py
@@ -6,7 +6,7 @@ import sys
 import pywt
 import matplotlib.figure
 
-from src.helpers import define_other_module_log_level
+from src.logging_helpers import define_other_module_log_level
 from src import retrieve_data, dwt, regression
 
 # * Logging settings
