+++ tests/test_retrieve_data.py
@@ -7,8 +7,8 @@ import sys
 import numpy as np
 import pandas as pd
 
-from src.helpers import define_other_module_log_level
-from src import ids
+from src.logging_helpers import define_other_module_log_level
+from constants import ids
 from src import retrieve_data
 
 # * Logging settings
