+++ tests/test_xwt.py
@@ -6,7 +6,7 @@ import sys
 
 import numpy as np
 
-from src.helpers import define_other_module_log_level
+from src.logging_helpers import define_other_module_log_level
 from src import retrieve_data
 from src import wavelet_helpers
 from src import xwt
